refactoring and fix bugs
