+++ app/frontend/pages/get_coordinates_page.py
@@ -2,7 +2,7 @@ import streamlit as st
 import requests
 from PIL import Image
 from loguru import logger
-from config import API_BASE_URL
+from settings.config import API_BASE_URL
 
 st.title("Get Coordinates from Image")
