+++ app/frontend/settings/config.py
@@ -1,6 +1,5 @@
 import os
 from dotenv import load_dotenv
-from loguru import logger
 
 
 environment = os.getenv("ENVIRONMENT", "development")
